Allow re-initialisation of initial thread TLS.
Due to the manner in which sel4test 'forks', it does not configure a
valid TLS for the new process. This change allows it to re-initialise
the TLS of the initial thread (assuming it has the same essential CNode
layout that the initial process did at startup).

This change is hazardous and makes it easier to configure an invalid TLS
and thread object within a process and should be used with care.