Fix alignment of entrypoints on x86_64.
This ensures that when C is entered into from either entry point, the
stack is aligned to 16 bytes + one word (for pushed return address) as
expected by gcc.

Additionally, when the _start or _sel4_start function calls into the C
runtime, the function should never return. An infinite loop guard is
added after the call to catch this case.