Add additional functions to aid rumprun.
This provides rumprun with enough information to allocate and
pre-initalise the TLS with minimal modification. This should also assist
in the creation of green-threading libraries.

Make the initialisation process clearer to better describe the resultant
TLS layout across differing platforms. Also fixes alignment issue on
ia32 when TLS is of an odd word size.