configurable exit callback
Allow the process to provide a callback to be used to exit the
applicaiton.

As this runtime is lightweight, it knows nothing about the environment
in which it is executing and so there is no way to exit the process
'properly'.

If a process attempts to exit and there is no configured exit callback,
the thread tha called exit will fault after all destructors have run.

Signed-off-by: Curtis Millar <curtis@curtism.me>