Don't set external __sysinfo in runtime
This removes the assumption in the runtime that a __vsyscall references
to a virtual syscall handler must be set by the runtime in an external
libc.

To ensure that musl can still initialise, there is a reference made to
the vsyscall handler in the stack of the root task if one exists
(otherwise the stack references a dummy).

Signed-off-by: Curtis Millar <curtis@curtism.me>