Force stack order for system-V info
The order, spacing and alignment of the argv, envp, and auxv regions are
part of the system-V specification and the init in muslc uses this to
derive the location of auxv from envp.

This uses an anonymous struct to enforce the ordering and spacing of
these items in release builds such that muslc can still initialise libc.